PredCoAPI/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.28.24
botocore==1.31.24
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
firebase-admin==6.3.0
frozenlist==1.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
kombu==5.3.1
msgpack==1.0.7
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
paho-mqtt==1.6.1
pandas==2.1.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-nginx==1.5.7
python3-openid==3.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sendgrid==6.10.0
shapely==2.0.2
sib-api-v3-sdk==7.6.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
twilio==8.5.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
yarl==1.9.2
